# backend/openapi-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TodoCamisetas API",
        "description": "API REST para la gestión de camisetas deportivas, clientes B2B, tallas y ventas. Permite administrar el stock de camisetas, gestionar clientes preferenciales y regulares, calcular precios con descuentos y registrar ventas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
...
        "/ventas": {
            "get": {
                "tags": [
                    "Ventas"
                ],
                "summary": "Listar ventas",
                "description": "Obtiene todas las ventas registradas junto con sus clientes y detalles.",
                "operationId": "listarVentas",
                "responses": {
                    "200": {
                        "description": "Listado de ventas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Venta"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Ventas"
                ],
                "summary": "Registrar venta",
                "description": "Registra una venta junto con sus detalles. El sistema valida stock disponible y descuenta automáticamente las unidades vendidas.",
                "operationId": "crearVenta",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VentaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Venta registrada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Venta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos o stock insuficiente"
                    },
                    "500": {
                        "description": "Error interno al registrar la venta"
                    }
                }
            }
        },
        "/ventas/{venta}": {
            "get": {
                "tags": [
                    "Ventas"
                ],
                "summary": "Obtener venta",
                "description": "Obtiene una venta específica junto con su cliente y detalle de productos.",
                "operationId": "obtenerVenta",
                "parameters": [
                    {
                        "name": "venta",
                        "in": "path",
                        "description": "ID de la venta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Venta encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Venta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Ventas"
                ],
                "summary": "Actualizar venta",
                "description": "Actualiza los datos generales de una venta.",
                "operationId": "actualizarVenta",
                "parameters": [
                    {
                        "name": "venta",
                        "in": "path",
                        "description": "ID de la venta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fecha_venta": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Venta actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Venta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ventas"
                ],
                "summary": "Eliminar venta",
                "description": "Elimina una venta siempre que no posea detalles asociados.",
                "operationId": "eliminarVenta",
                "parameters": [
                    {
                        "name": "venta",
                        "in": "path",
                        "description": "ID de la venta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Venta eliminada correctamente"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Camiseta": {
                "title": "Camiseta",
                "description": "Modelo completo de camiseta deportiva",
                "required": [
                    "titulo",
                    "club",
                    "pais",
                    "tipo",
                    "color",
                    "precio",
                    "stock",
                    "codigo_producto"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "titulo": {
                        "type": "string",
                        "example": "Camiseta Local 2025 - Selección Chilena"
                    },
                    "club": {
                        "type": "string",
                        "example": "Selección Chilena"
                    },
                    "pais": {
                        "type": "string",
                        "example": "Chile"
                    },
                    "tipo": {
                        "type": "string",
                        "example": "Local"
                    },
                    "color": {
                        "type": "string",
                        "example": "Rojo"
                    },
                    "precio": {
                        "type": "integer",
                        "example": 45000
                    },
                    "precio_final": {
                        "description": "Precio calculado según el descuento aplicado al cliente consultante",
                        "type": "integer",
                        "example": 40500,
                        "nullable": true
                    },
                    "stock": {
                        "type": "integer",
                        "example": 20
                    },
                    "detalles": {
                        "type": "string",
                        "example": "Edición aniversario 2025",
                        "nullable": true
                    },
                    "codigo_producto": {
                        "type": "string",
                        "example": "SCL2025L"
                    },
                    "tallas": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Talla"
                        },
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Cliente": {
                "title": "Cliente",
                "description": "Modelo completo de cliente B2B",
                "required": [
                    "nombre_comercial",
                    "rut_comercial",
                    "direccion",
                    "categoria",
                    "contacto_nombre",
                    "contacto_email",
                    "porcentaje_oferta"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre_comercial": {
                        "type": "string",
                        "example": "90minutos"
                    },
                    "rut_comercial": {
                        "type": "string",
                        "example": "76123456-7"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Providencia, Santiago"
                    },
                    "categoria": {
                        "type": "string",
                        "example": "Preferencial",
                        "enum": [
                            "Regular",
                            "Preferencial"
                        ]
                    },
                    "contacto_nombre": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "contacto_email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@90minutos.cl"
                    },
                    "porcentaje_oferta": {
                        "type": "integer",
                        "example": 10
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "DetalleVenta": {
                "title": "Detalle de Venta",
                "description": "Detalle asociado a una venta",
                "required": [
                    "venta_id",
                    "camiseta_id",
                    "cantidad",
                    "precio_unitario",
                    "subtotal"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "venta_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "camiseta_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cantidad": {
                        "type": "integer",
                        "example": 2
                    },
                    "precio_unitario": {
                        "type": "integer",
                        "example": 40500
                    },
                    "subtotal": {
                        "type": "integer",
                        "example": 81000
                    },
                    "camiseta": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Camiseta"
                            }
                        ],
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "DetalleVentaInput": {
                "title": "Detalle de Venta Input",
                "description": "Datos necesarios para agregar una camiseta a una venta",
                "required": [
                    "camiseta_id",
                    "cantidad"
                ],
                "properties": {
                    "camiseta_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cantidad": {
                        "type": "integer",
                        "example": 2,
                        "minimum": 1
                    }
                },
                "type": "object"
            },
            "Talla": {
                "title": "Talla",
                "description": "Modelo completo de talla disponible para camisetas",
                "required": [
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "M"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Venta": {
                "title": "Venta",
                "description": "Modelo completo de una venta realizada por un cliente",
                "required": [
                    "cliente_id",
                    "fecha_venta"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cliente_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "fecha_venta": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-06-10"
                    },
                    "cliente": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Cliente"
                            }
                        ],
                        "nullable": true
                    },
                    "detalles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DetalleVenta"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "VentaInput": {
                "title": "Venta Input",
                "description": "Datos necesarios para registrar una venta",
                "required": [
                    "cliente_id",
                    "fecha_venta",
                    "detalles"
                ],
                "properties": {
                    "cliente_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "fecha_venta": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-06-10"
                    },
                    "detalles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DetalleVentaInput"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}